Fix some warning for windows in string_conv.

+++ bench/aligned_buffer/aligned_buffer.cpp
@@ -9,7 +9,7 @@ static void fst_bench_array_loop(benchmark::State& state) {
   for (auto _ : state) {
     std::array<float, 512> buffer;
     for (std::size_t i = 0; i < buffer.size(); i++) {
-      buffer[i] = i;
+      buffer[i] = (float)i;
     }
     benchmark::ClobberMemory();
   }
@@ -20,7 +20,7 @@ static void fst_bench_aligned_buffer_loop(benchmark::State& state) {
   for (auto _ : state) {
     fst::stack_buffer<float, 512> buffer;
     for (std::size_t i = 0; i < buffer.size(); i++) {
-      buffer[i] = i;
+      buffer[i] = (float)i;
     }
     benchmark::ClobberMemory();
   }
@@ -31,8 +31,8 @@ static void fst_bench_struct_array_loop(benchmark::State& state) {
   for (auto _ : state) {
     std::array<std::pair<int, int>, 512> buffer;
     for (std::size_t i = 0; i < buffer.size(); i++) {
-      buffer[i].first = i;
-      buffer[i].second = i * 8;
+      buffer[i].first = (int)i;
+      buffer[i].second = (int)i * 8;
     }
     benchmark::ClobberMemory();
   }
@@ -43,8 +43,8 @@ static void fst_bench_struct_aligned_buffer_loop(benchmark::State& state) {
   for (auto _ : state) {
     fst::stack_buffer<std::pair<int, int>, 512> buffer;
     for (std::size_t i = 0; i < buffer.size(); i++) {
-      buffer[i].first = i;
-      buffer[i].second = i * 8;
+      buffer[i].first = (int)i;
+      buffer[i].second = (int)i * 8;
     }
     benchmark::ClobberMemory();
   }
@@ -55,7 +55,7 @@ static void fst_bench_vector_loop(benchmark::State& state) {
   for (auto _ : state) {
     std::vector<float> buffer(512);
     for (std::size_t i = 0; i < buffer.size(); i++) {
-      buffer[i] = i;
+      buffer[i] = (float)i;
     }
     benchmark::ClobberMemory();
   }
@@ -66,7 +66,7 @@ static void fst_bench_heap_aligned_buffer_loop(benchmark::State& state) {
   for (auto _ : state) {
     fst::heap_buffer<float, 512> buffer;
     for (std::size_t i = 0; i < buffer.size(); i++) {
-      buffer[i] = i;
+      buffer[i] = (float)i;
     }
     benchmark::ClobberMemory();
   }
